`script_from_string` as test class method
And move that test class into the `interactive_scripts` module to be subclassed
